django-gunicorn-nginx/rest_server/requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.6
django-rest-swagger==2.2.0
djangorestframework==3.11.0
gevent==20.6.0
greenlet==0.4.16
gunicorn==20.0.4
idna==2.9
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
openapi-codec==1.3.2
pytz==2020.1
requests==2.23.0
simplejson==3.17.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.9
zope.event==4.4
zope.interface==5.1.0
drf-yasg==1.17.1
